requirements.txt
tensorflow-intel
numpy
mlflow
flask

